state dla widoku
