+++ app/src/main/java/com/example/chessclockk/screens/settings/SettingsScreen.kt
@@ -1,11 +1,6 @@
-package com.example.chessclockk.views
+package com.example.chessclockk.screens.settings
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.terrakok.modo.Screen
