+++ app/src/main/java/com/example/chessclockk/vm/ClockState.kt
@@ -1,0 +1,6 @@
+package com.example.chessclockk.vm
+
+data class ClockState(
+    val whiteClock: String,
+    val blackClock: String
+)
