+++ app/src/main/java/com/example/chessclockk/vm/GameState.kt
@@ -1,4 +1,4 @@
-package com.example.chessclockk
+package com.example.chessclockk.vm
 
 enum class GameState {
     WHITE_MOVE, BLACK_MOVE, PAUSE, NEW_GAME
