+++ app/src/main/java/com/example/chessclockk/vm/IMainActivityVM.kt
@@ -1,0 +1,11 @@
+package com.example.chessclockk.vm
+
+interface IMainActivityVM {
+
+    fun onClockBlackPressed()
+    fun onClockWhitePressed()
+    fun onRestartClicked()
+    fun onRestartConfirmedClicked()
+    fun onPlayPauseBtnClicked()
+    fun onCustomTimeSet(customTime: String, bonus: String)
+}
