+++ app/src/main/java/com/example/chessclockk/vm/MainScreenState.kt
@@ -1,0 +1,8 @@
+package com.example.chessclockk.vm
+
+data class MainScreenState(
+    val timeFormat: String,
+    val gameState: GameState,
+    val blackMovesCount: Int,
+    val whiteMovesCount: Int,
+)
